Updated some variables to double

+++ Prac-02/C/src/CHeterodyning.c
@@ -3,11 +3,11 @@
 extern double data [SAMPLE_COUNT];
 extern double carrier[SAMPLE_COUNT];
 
-float result [SAMPLE_COUNT];
+double result [SAMPLE_COUNT];
 
 int main(int argc, char**argv){
     printf("Running Unthreaded Test\n");
-    printf("Precision sizeof %ld\n", sizeof(float));
+    printf("Precision sizeof %ld\n", sizeof(double));
     
 
     printf("Total amount of samples: %ld\n", sizeof(data) / sizeof(data[0]));
